requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
requests>=2.31.0
google-genai>=1.0.0
